+++ .typos.toml
@@ -1,2 +1,3 @@
 [default.extend-words]
 ket = "ket"
+BA = "BA"
